+++ Example/Tests/BLEManagerTests.swift
@@ -6,7 +6,6 @@
 //  Copyright © 2016 Huf Secure Mobile GmbH. All rights reserved.
 //
 
-import CommonUtils
 import CoreBluetooth
 import CryptoSwift
 @testable import SecureAccessBLE