+++ SecureAccessBLE/Classes/Managers/ConnectionManager/PhysicalConnectionChange.swift
@@ -6,7 +6,6 @@
 //  Copyright © 2017 Huf Secure Mobile GmbH. All rights reserved.
 //
 
-import CommonUtils
 import Foundation
 
 struct PhysicalConnectionChange: ChangeType {