+++ SecureAccessBLE/Classes/Managers/SorcManager/ConnectionChange.swift
@@ -6,8 +6,6 @@
 //  Copyright © 2017 Huf Secure Mobile GmbH. All rights reserved.
 //
 
-import CommonUtils
-
 /// Describes a change of connection state
 public struct ConnectionChange: ChangeType, Equatable {
     /// The state the connection can be in