+++ SecureAccessBLE/Classes/Managers/SorcManager/DiscoveryChange.swift
@@ -6,8 +6,6 @@
 //  Copyright © 2017 Huf Secure Mobile GmbH. All rights reserved.
 //
 
-import CommonUtils
-
 /// A change (state and last action) that describes the discovery transitions
 public struct DiscoveryChange: ChangeType {
     /// The state a `DiscoveryChange` can be in